move get_cudnn_version from common to pytorch utils
